now we can open SurahView.kt at a different starting aya

+++ navigation/src/commonMain/kotlin/com/islamversity/navigation/model/SurahLocalModel.kt
@@ -7,7 +7,7 @@ import okio.BufferedSource
 data class SurahLocalModel(
     val surahID: String,
     val surahName : String,
-    val startingAyaOrder : Int,
+    val startingAyaOrder : Long,
 ){
     companion object Sinker : SinkSerializer<SurahLocalModel> {
         const val EXTRA_SURAH_DETAIL = "extra_surah_detail"
@@ -16,13 +16,13 @@ data class SurahLocalModel(
             SurahLocalModel(
                 readUtf8(),
                 readUtf8(),
-                readInt(),
+                readLong(),
             )
 
         override fun BufferedSink.writeToSink(obj: SurahLocalModel) {
             writeUtf8(obj.surahID)
             writeUtf8(obj.surahName)
-            writeInt(obj.startingAyaOrder)
+            writeLong(obj.startingAyaOrder)
         }
     }
 }
